+++ CHANGES.rst
@@ -22,6 +22,11 @@ Developer changes
   been changed from ``url`` to ``binderhub_url`` in order to make it more
   explicit. The old configuration key (``url``) will be deprecated in
   version v0.4.0)
+* Support for generating JUnit XML summary files via the ``'junit'``
+  configuration value, which can be useful for building on CI services such as
+  CircleCI. See the related `CircleCI doc <https://circleci.com/docs/2.0/collect-test-data/#metadata-collection-in-custom-test-steps>`__
+  and `blog post <https://circleci.com/blog/how-to-output-junit-tests-through-circleci-2-0-for-expanded-insights/>`__.
+
 
 
 v0.2.0